feat: rewrite cjs to es modules
